changed resize observer, had issues with implementation need to work on
