+++ .eslintrc.json
@@ -15,7 +15,7 @@
 			"jsx": true
 		},
 		"project": "tsconfig.json",
-		"tsconfigRootDir": "./",
+		"tsconfigRootDir": ".",
 		"settings": {
 			"react": {
 				"version": "detect"
@@ -42,9 +42,6 @@
 		"plugin:import/recommended",
 		"plugin:import/typescript",
 		"prettier",
-		"prettier/prettier",
-		"prettier/react",
-		"prettier/@typescript-eslint",
 		"plugin:prettier/recommended"
 	],
 	// plugins bring in special rules, prettierrc rules you set
