+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://json.schemastore.org/prettierrc",
 	"parser": "typescript",
-	" editorconfig": true,
+	"editorconfig": true,
 	"bracketSameLine": true,
 	"arrowParens": "always",
 	"bracketSpacing": true,
