+++ src/pages/index.tsx
@@ -1,12 +1,9 @@
-import type { NextPage } from 'next';
-import Image from 'next/image';
-import Link from 'next/link';
-import { Fragment } from "react";
+import { Fragment } from 'react';
 
 import Meta from '@global/Meta';
 import XTerm from '@components/XTerm';
 
-import hero from "../public/brandon-mask.png";
+import hero from '../public/brandon-mask.png';
 
 export default function LandingPage() {
 	return (
@@ -15,4 +12,4 @@ export default function LandingPage() {
 			<XTerm />
 		</>
 	);
-};
+}
