+++ tsconfig.json
@@ -1,23 +1,23 @@
 {
   "compilerOptions": {
-	"target": "ES2020",
+    "target": "ES2020",
     "lib": [
       "dom",
       "dom.iterable",
       "esnext"
     ],
     "forceConsistentCasingInFileNames": true,
-	"noEmit": true,
+    "noEmit": true,
     "esModuleInterop": true,
-	"module": "ES2020",
+    "module": "ES2020",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-	"jsx": "react-jsx",
+    "jsx": "preserve",
     "incremental": true,
     "baseUrl": "./src",
-	"allowImportingTsExtensions": true,
-	"allowUmdGlobalAccess": true,
+    "allowImportingTsExtensions": true,
+    "allowUmdGlobalAccess": true,
     "paths": {
       "@components/*": [
         "components/*"
@@ -61,8 +61,8 @@
     "next-env.d.ts",
     "styled.d.ts",
     "next.config.js",
-	"src/**/*.tsx",
-	"src/**/*.ts",
+    "src/**/*.tsx",
+    "src/**/*.ts",
     "public/"
   ],
   "exclude": [
